requirements.txt
streamlit
easyocr
opencv-python-headless
numpy
pillow

